moved the data fetch hook into the lab folder for better placement
