+++ client/src/components/exercise/lab9/pages/hooks/useDataService.js
@@ -1,7 +1,7 @@
-import useLabRepair from "../../../body/Repair/hooks/useLabRepair";
-import { RepairService } from "../../../../services/lab9/RepairService";
-import { ExerciseService } from "../../../../services/lab9/ExerciseService";
-import { GAME_STATES } from "../../../../constants/lab9";
+import useLabRepair from "../../../../body/Repair/hooks/useLabRepair";
+import { RepairService } from "../../../../../services/lab9/RepairService";
+import { ExerciseService } from "../../../../../services/lab9/ExerciseService";
+import { GAME_STATES } from "../../../../../constants/lab9";
 
 /**
  * usDataService(): is a custom hook to abstract the logic implementation for the
@@ -49,14 +49,17 @@ const useDataService = (user, section, defaultGameState) => {
   async function handleExerciseUpdate(body, section) {
     try {
       const { isComplete, userid } = body;
-      const { isAddressComplete, isDateComplete, isNavComplete } = await ExerciseService.fetchExercise({
-        userid: userid
-      });
+      const { isAddressComplete, isDateComplete, isNavComplete } =
+        await ExerciseService.fetchExercise({
+          userid: userid,
+        });
       if (isComplete) {
         const updatedBody = {
           userid: body.userid,
           isAddressComplete:
-            section === GAME_STATES.REPAIR_ADDRESS_FORM ? true : isAddressComplete,
+            section === GAME_STATES.REPAIR_ADDRESS_FORM
+              ? true
+              : isAddressComplete,
           isDateComplete:
             section === GAME_STATES.REPAIR_DATE_REPAIR ? true : isDateComplete,
           isNavComplete:
